requirements.txt
wheel
pyail
androguard
lmdb
lxml
asn1crypto
python-magic
google-play-scraper
flor
